xsettings-manager: added tentative HiDPI support.
This relies on Gdk/WindowScalingFactor, which currently doesn't work in Debian as the only version of cairo that support scaling is in their GIT repositories.
When the WindowScalingFactor is > 1, the DPI values are automatically adapted.

Also, Xft/DPI is not directly handled anymore, use the new setting 'dpi' instead.

Added settings:
window-scaling-factor
dpi

Removed settings:
xft-dpi (handled by the new setting 'dpi')