autostart/: parse OnlyShowIn and skip KDE autostart applications.
We do not have our own key to put in the OnlyShowIn, so we need to be able
to launch every application targeted to GTK+-based environments.

We currently only test for KDE, and if so, we simply skip the autostart of
the application.
In the future we should have some more robust tests, plus a whitelist so
that an ignored application could still be 'forced' to autostart.