DBusService: listen to Display()'s idle_changed signal and set session's IdleHint accordingly.
Now logind is aware when the system is in idle, and thus we can offload the idle action handling to it.
This is a good thing as the user can now select which action to take when the system is idle.
This, of course, "conflicts" with the xscreensaver timer, that needs to be disabled.