screenshot: make the screenshot service a separate standalone application.
The screenshot tool is now accessible from the vera-screenshot binary and it's not exported
to DBus anymore.
Unfortunately unresolvable issues with certain graphic cards made shooting multiple screenshots
not possible, as subsequent screenshots will be identical to the first.
Having the screenshot tool as as separate application ensures that the image will be freed
every time a screenshot is taken (because the app will exit), thus resolving this bug.

Signed-off-by: Eugenio Paolantonio (g7) <me@medesimo.eu>